Moving theme management to Layout component
