+++ src/components/App.js
@@ -1,20 +1,16 @@
-import {createContext, useState} from "react";
+import {useContext, useState} from "react";
 import Speakers from "./Speakers";
 import Header from "./Header";
-
-export const ThemeContext = createContext();
+import Layout from "./Layout";
 
 function App({ speakers }) {
-    const[theme, setTheme] = useState("light")
-
     return (
-        <ThemeContext.Provider value={{ setTheme, theme }}>
-            <div className={theme === "light" ?
-                "container-fluid light" : "container-fluid dark"}>
+        <Layout initialTheme="light">
+            <div>
                 <Header />
                 <Speakers />
             </div>
-        </ThemeContext.Provider>
+        </Layout>
     );
 }
