+++ src/components/Header.js
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { ThemeContext } from "./App";
+import { ThemeContext } from "./Layout";
 
 function Header() {
